refactor: jus-79 사용안하는 dto 제거, import 최적화
