+++ src/main/java/com/jyp/justplan/domain/place/GooglePlacesProperties.java
@@ -1,9 +1,9 @@
 package com.jyp.justplan.domain.place;
 
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
 
 @Component
 @Getter
