+++ src/main/java/com/jyp/justplan/domain/place/application/PlaceService.java
@@ -2,11 +2,17 @@ package com.jyp.justplan.domain.place.application;
 
 import com.jyp.justplan.domain.city.domain.City;
 import com.jyp.justplan.domain.place.GooglePlacesProperties;
-import com.jyp.justplan.domain.place.domain.*;
+import com.jyp.justplan.domain.place.domain.GoogleMapType;
+import com.jyp.justplan.domain.place.domain.GooglePlace;
+import com.jyp.justplan.domain.place.domain.GooglePlaceRepository;
+import com.jyp.justplan.domain.place.domain.Memo;
+import com.jyp.justplan.domain.place.domain.MemoRepository;
+import com.jyp.justplan.domain.place.domain.Place;
+import com.jyp.justplan.domain.place.domain.PlaceRepository;
 import com.jyp.justplan.domain.place.dto.request.MemoUpdateDto;
 import com.jyp.justplan.domain.place.dto.request.PlaceListRequest;
-import com.jyp.justplan.domain.place.dto.request.PlaceRequest;
 import com.jyp.justplan.domain.place.dto.request.PlacePlanUpdateDto;
+import com.jyp.justplan.domain.place.dto.request.PlaceRequest;
 import com.jyp.justplan.domain.place.dto.request.PlaceUpdateRequest;
 import com.jyp.justplan.domain.place.dto.response.GooglePlacesSearchResponse;
 import com.jyp.justplan.domain.place.dto.response.PlaceDetailResponse;
@@ -21,7 +27,6 @@ import com.jyp.justplan.domain.plan.exception.NoSuchPlanException;
 import com.jyp.justplan.domain.plan.exception.NoSuchUserPlanException;
 import com.jyp.justplan.domain.user.domain.User;
 import com.jyp.justplan.domain.user.domain.UserRepository;
-import java.net.URI;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.LinkedHashMap;
@@ -29,14 +34,11 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.util.UriComponentsBuilder;
-import reactor.core.Disposable;
 import reactor.core.publisher.Mono;
 
 @Service
