+++ src/main/java/com/jyp/justplan/domain/place/domain/Memo.java
@@ -2,12 +2,16 @@ package com.jyp.justplan.domain.place.domain;
 
 import com.jyp.justplan.domain.BaseEntity;
 import com.jyp.justplan.domain.place.dto.request.MemoUpdateDto;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
-import javax.persistence.*;
 
 @Builder(toBuilder = true)
 @Getter
