+++ src/main/java/com/jyp/justplan/domain/place/domain/PlaceRepository.java
@@ -2,7 +2,6 @@ package com.jyp.justplan.domain.place.domain;
 
 import java.util.List;
 import java.util.Optional;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlaceRepository extends JpaRepository<Place,Long> {
