+++ src/main/java/com/jyp/justplan/domain/place/dto/request/PlacePlanUpdateDto.java
@@ -1,10 +1,10 @@
 package com.jyp.justplan.domain.place.dto.request;
 
+import java.util.List;
+import java.util.Map;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import java.util.List;
-import java.util.Map;
 
 @AllArgsConstructor
 @NoArgsConstructor
