+++ src/main/java/com/jyp/justplan/domain/place/dto/request/PlaceRequest.java
@@ -1,7 +1,5 @@
 package com.jyp.justplan.domain.place.dto.request;
 
-import com.jyp.justplan.domain.place.domain.Place;
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
