--- src/main/java/com/jyp/justplan/domain/place/dto/response/GooglePlaceGeometry.java
@@ -1,14 +1,0 @@
-package com.jyp.justplan.domain.place.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class GooglePlaceGeometry {
-
-    private GooglePlaceLocation location;
-
-}
